fix(Settings): silence warning

+++ Blockchain/SettingsTableViewController.m
@@ -1205,7 +1205,7 @@ const int aboutPrivacyPolicy = 2;
 
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath
 {
-    NSString *cellIdentifier = [NSString stringWithFormat:@"s%li-r%li", indexPath.section, (long)indexPath.row];
+    NSString *cellIdentifier = [NSString stringWithFormat:@"s%li-r%li", (long)indexPath.section, (long)indexPath.row];
     UITableViewCell *cell = [tableView dequeueReusableCellWithIdentifier:cellIdentifier];
     
     if (cell == nil) {
